requirements.txt
asgiref==3.8.1
attrs==24.2.0
autobahn==24.4.2
Automat==24.8.1
Babel==2.8.0
bcrypt==3.2.0
blinker==1.4
certifi==2020.6.20
cffi==1.17.1
chardet==4.0.0
charset-normalizer==3.4.2
click==8.2.0
cloud-init==24.4.1
colorama==0.4.4
command-not-found==0.3
configobj==5.0.6
constantly==23.10.4
cryptography==44.0.0
daphne==4.1.2
dbus-python==1.2.18
distlib==0.3.7
distro==1.7.0
distro-info==1.1+ubuntu0.2
Django==4.2.11
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
enum34==1.1.10
envi==0.2.2
filelock==3.12.2
frappe-bench==5.25.1
gitdb==4.0.12
GitPython==3.1.44
honcho==2.0.0
httplib2==0.20.2
hyperlink==21.0.0
idna==3.10
importlib-metadata==4.6.4
incremental==24.7.2
jeepney==0.7.1
Jinja2==3.1.6
jsonpatch==1.32
jsonpointer==2.0
jsonschema==3.2.0
keyring==23.5.0
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
MarkupSafe==2.0.1
more-itertools==8.10.0
mysqlclient==2.2.1
netifaces==0.11.0
oauthlib==3.2.0
platformdirs==3.10.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pycurl==7.44.1
PyGObject==3.42.1
PyHamcrest==2.0.2
PyJWT==2.3.0
PyMySQL==1.1.1
pyOpenSSL==24.3.0
pyparsing==2.4.7
pyrsistent==0.18.1
pyserial==3.5
python-apt==2.4.0+ubuntu4
python-crontab==2.6.0
python-dateutil==2.9.0.post0
pytz==2022.1
PyYAML==5.4.1
requests==2.32.3
SecretStorage==3.3.1
semantic-version==2.10.0
service-identity==24.2.0
six==1.16.0
smmap==5.0.2
sqlparse==0.5.3
systemd-python==234
tomli==2.2.1
Twisted==24.10.0
txaio==23.1.1
typing_extensions==4.12.2
ubuntu-pro-client==8001
ufw==0.36.1
unattended-upgrades==0.1
urllib3==1.26.5
virtualenv==20.24.2
wadllib==1.3.6
zipp==1.0.0
zope.interface==7.2
